+++ packages/theme/pigeonposse/examples/config.ts
@@ -1,5 +1,11 @@
 import { defineConfig } from '@dovenv/core'
 
-import { pigeonposseTheme } from '../src/main'
+import {
+	getWorkspaceConfig,
+	pigeonposseTheme,
+} from '../src/main'
 
-export default defineConfig( pigeonposseTheme( { lint: undefined } ) )
+export default defineConfig( pigeonposseTheme( {
+	lint : undefined,
+	core : await getWorkspaceConfig( import.meta.url, '../../../../' ),
+} ) )