+++ packages/theme/banda/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @dovenv/theme-banda
 
+## 1.1.2
+
+### Patch Changes
+
+- Extend pigeonposse theme, fix issues from docs
+
+- Updated dependencies []:
+  - @dovenv/docs@1.1.2
+  - @dovenv/core@1.1.2
+  - @dovenv/convert@1.1.2
+  - @dovenv/examples@1.1.2
+  - @dovenv/lint@1.1.2
+  - @dovenv/repo@1.1.2
+  - @dovenv/templates@1.1.2
+  - @dovenv/todo@1.1.2
+  - @dovenv/workspace@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes