+++ packages/theme/pigeonposse/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @dovenv/theme-pigeonposse
 
+## 1.1.2
+
+### Patch Changes
+
+- Extend pigeonposse theme, fix issues from docs
+
+- Updated dependencies []:
+  - @dovenv/core@1.1.2
+  - @dovenv/theme-banda@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes